requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Caching==2.1.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23